stack limit of 50, not Infinity
We don't have infinite room to send data, we only have 100KB, even for
POST, so we need to be a little bit conservative. 50 should be enough
for most cases. If we need to handle more, we'll need to implement
compression on the client so our packet is less than 100KB like our
other server side clients.